+++ src/main/java/com/safeking/shop/global/exhandler/advice/ExControllerAdvice.java
@@ -2,7 +2,7 @@ package com.safeking.shop.global.exhandler.advice;
 
 import com.safeking.shop.global.Error;
 import com.safeking.shop.global.exception.MemberNotFoundException;
-import com.safeking.shop.global.jwt.exception.CacheException;
+import com.safeking.shop.global.exception.CacheException;
 import com.safeking.shop.global.jwt.exception.TokenNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -29,43 +29,34 @@ public class ExControllerAdvice {
         return ResponseEntity
                 .badRequest().body(new Error(100,e.getMessage()));
     }
-    @ExceptionHandler
-    public ResponseEntity<Error> usernameNotFoundExHandler(UsernameNotFoundException e){
-        log.error("[usernameNotFoundExHandler] ex",e);
 
-        return new ResponseEntity<>(
-                new Error(1100,e.getMessage()),HttpStatus.NOT_FOUND);
-    }
     @ExceptionHandler
     public ResponseEntity<Error> refreshTokenNotFoundExHandler(TokenNotFoundException e){
         log.error("[refreshTokenNotFoundExHandler] ex",e);
 
-        return new ResponseEntity<>(new Error(1200,e.getMessage()), HttpStatus.FORBIDDEN);
+        return new ResponseEntity<>(new Error(200,e.getMessage()), HttpStatus.FORBIDDEN);
     }
-    @ExceptionHandler
-    public ResponseEntity<Error> MemberNotFoundExceptionExHandler(MemberNotFoundException e){
-        log.error("[MemberNotFoundExceptionExHandler] ex",e);
 
-        return new ResponseEntity<>(new Error(1223,e.getMessage()), HttpStatus.NOT_FOUND);
-    }
-    @ExceptionHandler
-    public ResponseEntity<Error> EntityNotFoundExceptionExHandler(EntityNotFoundException e){
-        log.error("[EntityNotFoundException] ex",e);
-
-        return new ResponseEntity<>(new Error(5001,e.getMessage()), HttpStatus.NOT_FOUND);
-    }
     @ExceptionHandler
     public ResponseEntity<Error> CacheExceptionExHandler(CacheException e){
         log.error("[EntityNotFoundException] ex",e);
 
-        return new ResponseEntity<>(new Error(1998,e.getMessage()), HttpStatus.NOT_FOUND);
+        return new ResponseEntity<>(new Error(201,e.getMessage()), HttpStatus.NOT_FOUND);
     }
+
+    @ExceptionHandler
+    public ResponseEntity<Error> EntityNotFoundExceptionExHandler(EntityNotFoundException e){
+        log.error("[EntityNotFoundException] ex",e);
+
+        return new ResponseEntity<>(new Error(300,e.getMessage()), HttpStatus.NOT_FOUND);
+    }
+
 
     @ExceptionHandler
     public ResponseEntity<Error> EntityExistsExceptionExHandler(EntityExistsException e){
         log.error("[EntityExistsException] ex",e);
 
-        return new ResponseEntity<>(new Error(5002,e.getMessage()), HttpStatus.BAD_REQUEST);
+        return new ResponseEntity<>(new Error(301,e.getMessage()), HttpStatus.BAD_REQUEST);
     }
 
     @ExceptionHandler
@@ -84,7 +75,7 @@ public class ExControllerAdvice {
         }
 
         return new ResponseEntity<>(
-                new Error(300,builder.toString()),HttpStatus.BAD_REQUEST);
+                new Error(400,builder.toString()),HttpStatus.BAD_REQUEST);
     }
 
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
@@ -93,6 +84,21 @@ public class ExControllerAdvice {
         log.error("[exceptionHandler] ex",e);
 
         return new Error(999,e.getMessage());
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<Error> usernameNotFoundExHandler(UsernameNotFoundException e){
+        log.error("[usernameNotFoundExHandler] ex",e);
+
+        return new ResponseEntity<>(
+                new Error(1100,e.getMessage()),HttpStatus.NOT_FOUND);
+    }
+
+    @ExceptionHandler
+    public ResponseEntity<Error> MemberNotFoundExceptionExHandler(MemberNotFoundException e){
+        log.error("[MemberNotFoundExceptionExHandler] ex",e);
+
+        return new ResponseEntity<>(new Error(1200,e.getMessage()), HttpStatus.NOT_FOUND);
     }
 
 }
