+++ src/main/java/com/safeking/shop/global/jwt/filter/JwtAuthorizationFilter.java
@@ -6,8 +6,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.safeking.shop.domain.user.domain.repository.CacheMemberRepository;
 import com.safeking.shop.global.auth.PrincipalDetails;
 import com.safeking.shop.domain.user.domain.entity.member.Member;
-import com.safeking.shop.global.exception.MemberNotFoundException;
-import com.safeking.shop.global.jwt.exception.CacheException;
+import com.safeking.shop.global.exception.CacheException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
